test: expand critical e2e coverage and reorganize by screen
