+++ osouji-touban-service/docs/issue-20260328-reflection.md
@@ -1,0 +1,96 @@
+# 1. 概要
+
+今回の作業では、従来の dashboard 中心だった E2E テストを、業務上の影響が大きいミッションクリティカル画面まで拡張した。
+あわせて `tests/e2e` 配下の構成を画面単位に再編し、機能追加時に対象シナリオを見つけやすく、責務が分かりやすいテスト構造へ移行した。
+
+主目的は以下の2点。
+
+- 重要オペレーション（施設管理・ユーザー管理・掃除エリア管理・清掃計画）の回帰検知力を上げる
+- テスト資産を画面別に整理し、保守性と拡張性を高める
+
+# 2. 実装内容
+
+## 構成再編
+
+- 旧ファイルを削除
+  - `tests/e2e/dashboard.spec.ts`
+  - `tests/e2e/mobile-table-scroll.spec.ts`
+- 画面別ファイルを新規追加
+  - `tests/e2e/dashboard/dashboard.spec.ts`
+  - `tests/e2e/facilities/facilities.spec.ts`
+  - `tests/e2e/users/users.spec.ts`
+  - `tests/e2e/cleaning-areas/cleaning-areas.spec.ts`
+  - `tests/e2e/weekly-duty-plans/weekly-duty-plans.spec.ts`
+
+## 追加・維持した主要フロー
+
+### dashboard
+
+- 既存シナリオ（選択エリアの保持と当番表示）を維持
+- 既存価値を落とさず、他画面への拡張を優先できるように整理
+
+### facilities
+
+- フィルタ（状態）更新を連続実行しても `scrollY` が維持されることを検証
+- モーダルからの施設作成（入力・保存・完了トースト・一覧反映）を検証
+
+### users
+
+- フィルタ（状態）更新時の `scrollY` 維持を検証
+- モーダルからのユーザー作成（入力・保存・完了トースト・一覧反映）を検証
+
+### cleaning-areas
+
+- 施設/所属ユーザーフィルタ更新時の `scrollY` 維持を検証
+- モバイル表示で会員カードが表示されること、および「解除」操作が機能することを検証
+- デスクトップ表示でユーザーアサイン操作ができることを検証
+- 画面幅に応じた表示切り替え（カード/テーブル）を検証
+
+### weekly-duty-plans
+
+- モバイル幅（複数幅）でテーブル横スクロールが成立し、ページ全体の横オーバーフローが発生しないことを検証
+- フィルタ更新後の `scrollY` 維持を検証
+- 計画作成から発行までの一連フロー（成功トースト、発行ボタン状態遷移）を検証
+
+# 3. レビュー指摘と対応
+
+## 指摘1: `scrollY` 検証の偽陽性
+
+- 課題
+  - 更新前後の比較基準が曖昧なケースがあり、意図せず通る（または落ちる）可能性があった
+- 対応
+  - baseline 差分ベースの比較に統一し、更新直前の `scrollY` を基準に誤差許容内で判定する方式へ修正
+  - 連続更新ケースでも「直前の実測値」を次の baseline として引き継ぐ形に整理
+- 効果
+  - スクロール維持の回帰検知が安定し、偽陽性/偽陰性の双方を抑制
+
+## 指摘2: mutation モックの `If-Match` 検証不足
+
+- 課題
+  - 更新系APIモックがヘッダ前提を十分に検証しておらず、実運用に近い前提を満たせていなかった
+- 対応
+  - mutation モックで `If-Match` ヘッダ値をチェックし、不一致時は `412 Precondition Failed` を返す分岐を追加
+  - 該当フロー（メンバー割当/解除、計画発行）で楽観ロック前提をテスト側でも担保
+- 効果
+  - 競合制御を含むAPI契約の検証密度が向上し、E2Eの信頼性が上がった
+
+# 4. テスト結果
+
+- 実行コマンド: `npm run e2e`
+- 結果: `13 passed`
+
+画面別再編後のシナリオ群がすべて通過し、回帰なく移行できたことを確認した。
+
+# 5. 学びと次アクション（短め）
+
+## 学び
+
+- 画面別に責務を切ることで、テストの意図と失敗時の原因箇所が把握しやすくなる
+- `scrollY` のような UI 安定性検証は、比較基準（baseline）を明確化すると一気に信頼性が上がる
+- ETag/`If-Match` をモックでも厳密化すると、実装契約のズレを早期に発見できる
+
+## 次アクション
+
+- 主要 mutation 系シナリオに `If-Match` 前提の検証を横展開する
+- 重要画面のフィルタ/ページング/再描画の安定性ケースを段階的に追加する
+- 失敗時の診断を速めるため、共通ヘルパー（モック・スクロール検証）の再利用を進める
